Semi-done checkpoints and attack
Semi-done checkpoints: the inventory items are removed when I die and go back to the last checkpoint, the enemy reappears as well. NEED TO FIX THAT.

Cherry bomb: the angle is wrong, and I think it might be because the character moves their hand causing the angle of throw of the cherry bomb to change. NEED TO FIX THAT.